requirements.txt
fastapi
uvicorn
python-dotenv
masumi
pydantic
requests
httpx 